@NewFeature: website/utils.py ImportSupplierData, CampaignInventories - A new API to fetch all campaign inventories related details has been made in website/views. The api fetches all the campaiagn details + all shortlisted suppliers details + all shortlisted inventory details for each of the shortlisted suppliers.
The table ShortlistedInventoryPricingDetails is being modified to store campaign_status and phase. It also stores R.D, C.D.
This table stores all the inventories ids ever assigned to any proposal and can be queried for. This table has an Fk ko SS table which
signifies which supplier and under which proposal was this inventory shortlisted.

A new algorithm to shortlisted stall inventory is made and implemenetd when import-supplier-data is called. This algorithm ensures that
within a single proposal, and within same supplier, no two inventory ids are assigned to the proposal. A invenotory can be assigned to a
proposal under one time frame which is same as proposals start and end date. It can be reassigned to a different proposal under different
time frame which does not intersect with any already assigned time frames for that inventory_id. For more detailed explanations, look
at the comments in the code.

Finally an Audit table is made which stores audit dates for each  inventory_id.